Move export status constants to apps
So that it can be shared between sub apps
