+++ test/selectors/interaction-form.js
@@ -1,5 +1,5 @@
 const typeaheadId = '#group-field-dit_participants'
-const { EXPORT_INTEREST_STATUS } = require('../../src/apps/interactions/constants')
+const { EXPORT_INTEREST_STATUS } = require('../../src/apps/constants')
 
 module.exports = {
   subject: '#field-subject',
